Changed the getRawMessageLength accessor of the Message subclasses to return an int instead of a long; wrote the getRawMessage() accessor implementations for LoginMessage and ErrorMessage, and partial implementations for the rest.

+++ message/Message.java
@@ -11,7 +11,7 @@ public abstract class Message
 
 public abstract MessageCode getMessageCode();
 
-public abstract long getRawMessageLength();
+public abstract int getRawMessageLength();
 
 public abstract ByteBuffer getRawMessage();
